Colmar
The very first version of the game. I just coded the basic sprite class, and also a basic scene class, which contains a list to handle sprites. There is also a very minimalist debug module, that only logs things when in debug mode. But this mode must be hand-defined in "mydebug.lua".