Colmar Server
This is the really non-sense basic server of the game. It is only here to prepare the implementation of a later server, but for now, it is just something that gets data and prints it, "interpreting" (splitting) the message. So I already planned a message pattern to be sent by the client that is defined in the client's "SocketConnection.lua".