NetGame - Wettolsheim
This patch is really important! As I couldn't find any love method that checks if the user presses the window's exit button, I remove the option to close a window by clicking on the exit button. The server now swaps the waiting clients in the main list if a connected client disconnects or deletes the client's connection if the client switch the server it tries to connect to.