NetGame - Walbach
Little town, little changes! In this version, I just added a new system that is a scene managing system. It means, that now, the game doesn't handle with a particular scene, but handles with a modifiable set of scenes (See "SceneManager.lua"). Also, we don't just create a Scene, but a daughter class of Scene. The scenes are user-implemented, but must implement Scene:update(dt) and Scene:draw() calling respectively Scene:updateScene(dt) and Scene:drawScene() that for now only draw the sprites.